+++ pkg/template/master_template.go
@@ -323,6 +323,8 @@ systemd:
         {{ end -}}
         --node-ip=${DEFAULT_IPV4} \
         --config=/etc/kubernetes/config/kubelet.yaml \
+        --container-runtime=remote \
+        --container-runtime-endpoint=unix:///run/containerd/containerd.sock \
         --logtostderr=true \
         --cloud-provider={{.Cluster.Kubernetes.CloudProvider}} \
         --pod-infra-container-image={{ .Images.Pause }} \