+++ pkg/template/worker_template.go
@@ -223,6 +223,8 @@ systemd:
         {{ end -}}
         --node-ip=${DEFAULT_IPV4} \
         --config=/etc/kubernetes/config/kubelet.yaml \
+        --container-runtime=remote \
+        --container-runtime-endpoint=unix:///run/containerd/containerd.sock \
         --logtostderr=true \
         --cloud-provider={{.Cluster.Kubernetes.CloudProvider}} \
         --pod-infra-container-image={{ .Images.Pause }} \
@@ -406,6 +408,26 @@ storage:
       contents:
         source: "data:text/plain;charset=utf-8;base64,{{  index .Files "conf/ip_vs.conf" }}"
 
+    - path : /etc/containerd/config.toml
+      filesystem: root
+      mode: 420
+      user:
+        id: 0
+      group:
+        id: 0
+      contents:
+        source: "data:text/plain;charset=utf-8;base64,{{ index .Files "conf/containerd-config.toml" }}"
+ 
+    - path : /etc/systemd/system/containerd.service.d/10-use-custom-config.conf
+      filesystem: root
+      mode: 420
+      user:
+        id: 0
+      group:
+        id: 0
+      contents:
+        source: "data:text/plain;charset=utf-8;base64,{{ index .Files "conf/10-use-custom-config.conf" }}"
+ 
     - path: /opt/k8s-extract
       filesystem: root
       mode: 0544